Fix "_sync_metadata_ports" with no DHCP subnets
When a subnet does not have DHCP configured, the metadata port does
not have an IP address on this CIDR. The method
"OvnNbSynchronizer.sync_networks_ports_and_dhcp_opts", was always
setting an IP address for the metadata ports, regardless of the subnet
configuration (with or without DHCP).

The method "_sync_metadata_ports", in charge of synchronizing the
metadata ports, now filters the subnets by the parameter "enable_dhcp".

In case of having a subnet with DHCP enabled, if the metadata port is
missing the subnet IP addresses, the method adds them.

In case of having a subnet without DHCP enabled, if the metadata port
has an IP address on the subnet, the method removes it.

NOTE: when moving this patch from Neutron to networking-ovn, [1] was
merged first. This patch implements the needed code to fix LP#1939726,
that was in the Neutron one [2]. That means the current patch only
includes the testing part (that should be included in the repository
to avoid regressions).

[1]https://review.opendev.org/c/openstack/networking-ovn/+/812714
[2]https://review.opendev.org/c/openstack/neutron/+/811996

Closes-Bug: #1939726

Neutron-master: https://review.opendev.org/c/openstack/neutron/+/811996

Change-Id: I09cc14dff6933aae63cbd43a29f9221f405ecede