[OVN] Prevent deleting the only IP of a router port
Neutron allows deleting the only IP of a router port but
the OVN NB DB doesn't, since it expects that the
network value of a port is greater than 0. This should
not be possible since it causes that the DB are not
perfectly sync.

It is needed to check BEFORE_UPDATE if the port
that will be updated is of type router owned and
if it will have an IP after the update. If not
an error needs to be raised.

Comparet to master patch, this has changed payload to kwargs
Conflicts:
    neutron/services/ovn_l3/plugin.py

From Neutron-master:
    https://review.opendev.org/c/openstack/neutron/+/817525

Closes-Bug: #1948457
(manually cherry picked from commit fa2179278f064f75e057f7020e4f0c34f358de1b)

Change-Id: I8c1a595ba226e251d549432a489687114543c2c0