[OVN] Check new added segments in OVN mech driver
Check new segments added to OVN mechanism driver. This mechanism
driver does not inherit from ``SimpleAgentMechanismDriverBase``
because OVN has no agents (NOTE 1). However OVN mechanism driver
requires an OVN controller service in each deployed chassis. This
is how OVN driver knows the hosts using this backend.

The segments, attached to an external network (connected to a
physical network), can be mapped to a host if ther is a chassis (OVN
controller agent) in this host.

NOTE 1: OVN provides agent API compatibility, presenting the
controller and the metadata services as agents. But OVN mechanism
driver has no control or provides any information (via RPC) to
those services.

NOTE 2: this patch is the complementary of [1]. The missing code in
[1], the OVN part, is implemented in this patch, that depends on
the Neutron one.

[1]https://review.opendev.org/c/openstack/neutron/+/826271

Depends-On: https://review.opendev.org/c/openstack/neutron/+/826271

Closes-Bug: #1949967
Change-Id: I570b2251da8a4a25d155ba838346b24afafd727f