Ensure no GARPs are sent for Load Balancer VIPs on tenant networks
When a loadbalancer is created in tenant network (VIP and members),
and that tenant network is connected to a router, which is connected
to the provider network, the ovn loadbalancer gets
associated to the ovn logical router. If the "router" option is used for
"nat-addresses" (as currently done), ovn-controller will send GARPs for
that VIP.

As there is nothing blocking different tenants in openstack to create a
subnet with the same CIDR and then a loadbalancer with the same VIP,
there may be several ovn-controllers generating GARPs on the provider
network for the same IP, each one with the MAC of the logical router
port belonging to each tenant.
This could be a problem for the physical network infrastructure.

This patch fixes it by setting a new option added in OVN[1], named
"exclude-lb-vips-from-garp", on the router gateway port, ensuring
no GARPs are sent for the load balancer VIPs.

[1] https://patchwork.ozlabs.org/project/ovn/patch/183edfc446633c5c38d7d7361089d34432c527dd.1645793899.git.lorenzo.bianconi@redhat.com/

Closes-Bug: #1964901

(manually cherry picked from commit 9d4e6edd5b4c92c88fbe8d06f6a176fae1b35f25)

Change-Id: Ifc2e5b2cc64c0b3beafd0a2e6eb9b9f248970cc5