Retry logical switch associations to load balancers
On load-balancer creation, or other operations related to association
of the load balancer to the logical router, all logical switches
associated with the logical router in the target network are also
associated with the load-balancer.  When the topology includes multiple
subnets, it may happen that the operation over the load-balancer match
in time with the removal of some of the subnets.

The creation of the load-balancer is a transactional atomic
process. This patch is splitting such transactions and retrying
in case of the above error. In case the attempts are exhausted and the
error remains, we evaluate command by command, in case LsLbAdd or
LsLbDel of an Ls associated to the Lr, we can omit the error and go
forward.

Add extra security check to function ovn_name (networking_ovn/common/utils.py)
just to not fail in some neutron code/test that are calling to this
function with a None as argument, a soon as possible a new patch will be
submitted removing those calls.

Closes-Bug: #1963921
(manually cherry picked from commit
1a62902b5601c764da659f13242131828b4078ca)

Upstream-Yoga: https://review.opendev.org/c/openstack/ovn-octavia-provider/+/833872
Upstream-Wallaby: https://review.opendev.org/c/openstack/ovn-octavia-provider/+/833882
Upstream-Xena: https://review.opendev.org/c/openstack/ovn-octavia-provider/+/833885
Upstream-Victoria: https://review.opendev.org/c/openstack/ovn-octavia-provider/+/833884
Upstream-Ussuri: https://review.opendev.org/c/openstack/ovn-octavia-provider/+/833995

Change-Id: I2c7e0c677f0687990bbd71b3f8d511051dbe0359