[OVN] Pin OVS/OVN versions in Octavia job
Octavia job is also broken during compilation,
Pin to working OVN/OVN versions as done with
the rally job[1].

This patch also updates irrelevant files to ignore changes
in all job config files except zuul.d/project.yaml(as
this contains the list of jobs to run not their definitions).
With this only the jobs whose definitions are updated(along with
it's child jobs) will be triggered because by default
match-on-config-updates is set to true for a job[1] and
that ignores files or irrelevant-files for a job inclusion.

[1] https://review.opendev.org/c/openstack/networking-ovn/+/839070
[2] https://zuul-ci.org/docs/zuul/latest/config/job.html#attr-job.match-on-config-updates

Related-Bug: #1967472
Change-Id: I62da6480843bd56302503d7f312d040dada6cd82