Avoid LB ERROR status on delete when LS/LR are already deleted
The delete operation of an LB lists all references to it (LS and
LR) and in a single transaction deletes them together with the
LB itself. If any of those deletions raise a RowNotFound exception,
the execution will report ERROR status.

This patch will reduce those situations, rechecking the command
that triggers the RowNotFound exception and passing in case a LS
or a LR was the culprit, deleting finally the LB and reporting as
a DELETED status. In other case, the exception is raised and
finally the LB will be reported as ERROR status.

Closes-Bug:#1990129

(manually cherry picked from commit f9ccb4c48627a200ef27dde60b01bccdccac4e45)

Change-Id: I85265a2ab5e75d244a6aa85263d951540540e123