Remove "networking-ovn-dsvm-functional-py27" job
This job is currently failing. Because this branch is in EM and we
don't have resources to fix this job now, we are removing it from
the gate and check queues.

This patch is closing this related bug. If in a future a new patch
is sent to fix the job, the bug will be reopened.

Closes-Bug: #1997262
Change-Id: Ifc4fdaf22f9768e4bffd5c3b4f2917a092859c4b